+++ module/webroot/scripts/file_selector.js
@@ -102,9 +102,9 @@ async function listFiles(path, skipAnimation = false) {
                             chmod 644 ${basePath}/custom_${fileName}
                         `);
                         closeFileSelector();
-                        showPrompt('global.saved', true, undefined, undefined, `${basePath}/custom_${fileName}`);
+                        showPrompt('global_saved', true, undefined, undefined, `${basePath}/custom_${fileName}`);
                     } catch (error) {
-                        showPrompt('global.save_fail', false);
+                        showPrompt('global_save_fail', false);
                     }
                 }
             });
@@ -214,14 +214,14 @@ export async function openFileSelector(type) {
             if (item && item.querySelector('span').textContent.endsWith('.json')) {
                 try {
                     const jsonConfig = await exec(`cat ${currentPath}/${item.querySelector('span').textContent}`);
-                    resolve(jsonConfig.trim());
                     closeFileSelector();
+                    resolve(jsonConfig.trim());
                 } catch (error) {
                     reject(error);
                 }
             } else if (item && item.querySelector('span').textContent.endsWith('.txt')) {
-                resolve(true);
                 closeFileSelector();
+                resolve(true);
             }
         });
     });